move assets to upper level
