+++ assets/assets.go
